Extract InferenceProfile operations into document_config/inference_profile.rs
