+++ crates/defra-agent/src/document_config/mod.rs
@@ -17,23 +17,18 @@ pub(crate) use behavior::{
 };
 use behavior::create_default_behavior;
 
+pub use inference_profile::{load_inference_profile, upsert_inference_profile, InferenceProfile};
+#[allow(unused_imports)]
+pub(crate) use inference_profile::{
+    list_inference_profile_records, load_inference_profile_by_doc_id,
+    load_inference_profile_record,
+};
+
 use anyhow::{anyhow, Result};
 use defra_node::EmbeddedNode;
 use serde::{Deserialize, Serialize};
 
 use crate::graphql::escape_graphql_string;
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct InferenceProfile {
-    pub profile_id: String,
-    pub display_name: Option<String>,
-    pub context_window: Option<i64>,
-    pub max_output_tokens: Option<i64>,
-    pub max_turns: Option<i64>,
-    pub temperature: Option<f64>,
-    pub stream_batch_ms: Option<i64>,
-    pub deadline_duration_secs: Option<i64>,
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct ToolSelectionDocument {
@@ -157,85 +152,6 @@ pub async fn ensure_agent_principal(
         created_principal,
         created_default_behavior,
     })
-}
-
-pub async fn load_inference_profile(
-    node: &EmbeddedNode,
-    profile_id: &str,
-) -> Result<Option<InferenceProfile>> {
-    Ok(load_inference_profile_record(node, profile_id)
-        .await?
-        .map(|(_, profile)| profile))
-}
-
-pub(crate) async fn load_inference_profile_record(
-    node: &EmbeddedNode,
-    profile_id: &str,
-) -> Result<Option<(String, InferenceProfile)>> {
-    let escaped_profile_id = escape_graphql_string(profile_id);
-    let query = format!(
-        r#"{{
-            InferenceProfile(
-                filter: {{ profile_id: {{ _eq: "{escaped_profile_id}" }} }},
-                limit: 1
-            ) {{
-                _docID
-                profile_id
-                display_name
-                context_window
-                max_output_tokens
-                max_turns
-                temperature
-                stream_batch_ms
-                deadline_duration_secs
-            }}
-        }}"#
-    );
-
-    let resp = node.execute(&query).await;
-    if resp.has_errors() {
-        anyhow::bail!("query InferenceProfile failed: {:?}", resp.errors);
-    }
-
-    Ok(serde_helpers::first_row_with_doc_id(
-        resp.data.as_ref(),
-        "InferenceProfile",
-    ))
-}
-
-pub(crate) async fn load_inference_profile_by_doc_id(
-    node: &EmbeddedNode,
-    doc_id: &str,
-) -> Result<Option<(String, InferenceProfile)>> {
-    let escaped_doc_id = escape_graphql_string(doc_id);
-    let query = format!(
-        r#"{{
-            InferenceProfile(
-                filter: {{ _docID: {{ _eq: "{escaped_doc_id}" }} }},
-                limit: 1
-            ) {{
-                _docID
-                profile_id
-                display_name
-                context_window
-                max_output_tokens
-                max_turns
-                temperature
-                stream_batch_ms
-                deadline_duration_secs
-            }}
-        }}"#
-    );
-
-    let resp = node.execute(&query).await;
-    if resp.has_errors() {
-        anyhow::bail!("query InferenceProfile by _docID failed: {:?}", resp.errors);
-    }
-
-    Ok(serde_helpers::first_row_with_doc_id(
-        resp.data.as_ref(),
-        "InferenceProfile",
-    ))
 }
 
 pub async fn load_tool_selection(
@@ -380,31 +296,6 @@ pub(crate) async fn list_all_tool_selection_records(
     Ok(serde_helpers::rows_with_doc_id(resp.data.as_ref(), "ToolSelection"))
 }
 
-pub(crate) async fn list_inference_profile_records(
-    node: &EmbeddedNode,
-) -> Result<Vec<(String, InferenceProfile)>> {
-    let query = r#"{
-            InferenceProfile(order: { profile_id: ASC }) {
-                _docID
-                profile_id
-                display_name
-                context_window
-                max_output_tokens
-                max_turns
-                temperature
-                stream_batch_ms
-                deadline_duration_secs
-            }
-        }"#;
-
-    let resp = node.execute(query).await;
-    if resp.has_errors() {
-        anyhow::bail!("list InferenceProfile failed: {:?}", resp.errors);
-    }
-
-    Ok(serde_helpers::rows_with_doc_id(resp.data.as_ref(), "InferenceProfile"))
-}
-
 pub async fn upsert_tool_selection(
     node: &EmbeddedNode,
     selection: &ToolSelectionDocument,
@@ -470,62 +361,6 @@ pub async fn upsert_tool_selection(
     let resp = node.execute(&mutation).await;
     if resp.has_errors() {
         anyhow::bail!("upsert ToolSelection failed: {:?}", resp.errors);
-    }
-    Ok(())
-}
-
-pub async fn upsert_inference_profile(
-    node: &EmbeddedNode,
-    profile: &InferenceProfile,
-) -> Result<()> {
-    let escaped_profile_id = escape_graphql_string(&profile.profile_id);
-
-    let add_fields = vec![
-        Some(format!(r#"profile_id: "{escaped_profile_id}""#)),
-        graphql_fields::graphql_string_field("display_name", profile.display_name.as_deref()),
-        graphql_fields::graphql_optional_int_field("context_window", profile.context_window),
-        graphql_fields::graphql_optional_int_field("max_output_tokens", profile.max_output_tokens),
-        graphql_fields::graphql_optional_int_field("max_turns", profile.max_turns),
-        graphql_fields::graphql_optional_float_field("temperature", profile.temperature),
-        graphql_fields::graphql_optional_int_field("stream_batch_ms", profile.stream_batch_ms),
-        graphql_fields::graphql_optional_int_field("deadline_duration_secs", profile.deadline_duration_secs),
-    ]
-    .into_iter()
-    .flatten()
-    .collect::<Vec<_>>()
-    .join(",\n                    ");
-
-    let update_fields = vec![
-        graphql_fields::graphql_string_field("display_name", profile.display_name.as_deref()),
-        graphql_fields::graphql_optional_int_field("context_window", profile.context_window),
-        graphql_fields::graphql_optional_int_field("max_output_tokens", profile.max_output_tokens),
-        graphql_fields::graphql_optional_int_field("max_turns", profile.max_turns),
-        graphql_fields::graphql_optional_float_field("temperature", profile.temperature),
-        graphql_fields::graphql_optional_int_field("stream_batch_ms", profile.stream_batch_ms),
-        graphql_fields::graphql_optional_int_field("deadline_duration_secs", profile.deadline_duration_secs),
-    ]
-    .into_iter()
-    .flatten()
-    .collect::<Vec<_>>()
-    .join(",\n                    ");
-
-    let mutation = format!(
-        r#"mutation {{
-            upsert_InferenceProfile(
-                filter: {{ profile_id: {{ _eq: "{escaped_profile_id}" }} }},
-                add: {{
-                    {add_fields}
-                }},
-                update: {{
-                    {update_fields}
-                }}
-            ) {{ _docID }}
-        }}"#
-    );
-
-    let resp = node.execute(&mutation).await;
-    if resp.has_errors() {
-        anyhow::bail!("upsert InferenceProfile failed: {:?}", resp.errors);
     }
     Ok(())
 }
